NPCファイルの命名規則を変更した
Normal、Hardなどのコース名をPrefixとする。
