+++ src/js/npc/hard-lightning-dozer.js
@@ -1,7 +1,7 @@
 // @flow
 
 import type {NPC} from "./npc";
-import {ArmDozerIdList, ArmDozers, PilotIds, Pilots} from "gbraver-burst-core";
+import {ArmDozerIdList, ArmDozers, correctPower, PilotIds, Pilots} from "gbraver-burst-core";
 import type {SimpleRoutine} from "./simple-npc";
 import {SimpleNPC} from "./simple-npc";
 
@@ -16,12 +16,17 @@ const ZERO_BATTERY = {
  * 攻撃ルーチン
  */
 const attackRoutine: SimpleRoutine = data => {
-  const burst = data.commands.find(v => v.type === 'BURST_COMMAND');
-  const allBattery = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === data.enemy.armdozer.battery);
+  const hasCorrectPower = 0 < correctPower(data.enemy.armdozer.effects);
+  const pilot = data.commands.find(v => v.type === 'PILOT_SKILL_COMMAND');
+  const battery5 = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === 5);
   const allBatteryMinusOne = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === data.enemy.armdozer.battery - 1);
 
-  if (burst && allBattery) {
-    return allBattery;
+  if (pilot) {
+    return pilot;
+  }
+
+  if (hasCorrectPower && battery5) {
+    return battery5;
   }
 
   if (allBatteryMinusOne) {
@@ -36,20 +41,10 @@ const attackRoutine: SimpleRoutine = data => {
  * 防御ルーチン
  */
 const defenseRoutine: SimpleRoutine = data => {
-  const pilot = data.commands.find(v => v.type === 'PILOT_SKILL_COMMAND');
   const burst = data.commands.find(v => v.type === 'BURST_COMMAND');
-  const battery5 = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === 5);
   const battery1 = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === 1);
 
-  if (pilot && battery5) {
-    return pilot;
-  }
-
-  if (burst && battery5) {
-    return battery5;
-  }
-
-  if (burst && data.enemy.armdozer.battery === 0) {
+  if (burst) {
     return burst;
   }
 
@@ -61,12 +56,12 @@ const defenseRoutine: SimpleRoutine = data => {
 };
 
 /**
- * パーファクト ネオランドーザ
+ * ハードコース ライトニングドーザ NPC
  *
- * @return NPC
+ * @returns NPC
  */
-export function perfectNeoLandozer(): NPC {
-  const armdozer = ArmDozers.find(v => v.id === ArmDozerIdList.NEO_LANDOZER) ?? ArmDozers[0];
-  const pilot = Pilots.find(v => v.id === PilotIds.TSUBASA) ?? Pilots[0];
+export function hardLightningDozer(): NPC {
+  const armdozer = ArmDozers.find(v => v.id === ArmDozerIdList.LIGHTNING_DOZER) ?? ArmDozers[0];
+  const pilot = Pilots.find(v => v.id === PilotIds.GAI) ?? Pilots[0];
   return new SimpleNPC(armdozer, pilot, attackRoutine, defenseRoutine);
 }
