+++ src/js/npc/hard-neo-landozer.js
@@ -1,10 +1,9 @@
 // @flow
 
 import type {NPC} from "./npc";
-import {ArmDozerIdList, ArmDozers, correctPower, PilotIds, Pilots} from "gbraver-burst-core";
+import {ArmDozerIdList, ArmDozers, PilotIds, Pilots} from "gbraver-burst-core";
 import type {SimpleRoutine} from "./simple-npc";
 import {SimpleNPC} from "./simple-npc";
-import {canBeatDown} from "./can-beat-down";
 
 /** 0バッテリー */
 const ZERO_BATTERY = {
@@ -20,14 +19,8 @@ const attackRoutine: SimpleRoutine = data => {
   const burst = data.commands.find(v => v.type === 'BURST_COMMAND');
   const allBattery = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === data.enemy.armdozer.battery);
   const allBatteryMinusOne = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === data.enemy.armdozer.battery - 1);
-  const hasPlusCorrectPower = 0 < correctPower(data.enemy.armdozer.effects);
-  const hasDamaged = data.enemy.armdozer.hp  < data.enemy.armdozer.maxHp;
 
-  if (hasDamaged && burst) {
-    return burst;
-  }
-
-  if (hasPlusCorrectPower && allBattery) {
+  if (burst && allBattery) {
     return allBattery;
   }
 
@@ -43,12 +36,21 @@ const attackRoutine: SimpleRoutine = data => {
  * 防御ルーチン
  */
 const defenseRoutine: SimpleRoutine = data => {
+  const pilot = data.commands.find(v => v.type === 'PILOT_SKILL_COMMAND');
   const burst = data.commands.find(v => v.type === 'BURST_COMMAND');
-  const allBattery = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === data.enemy.armdozer.battery);
+  const battery5 = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === 5);
   const battery1 = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === 1);
 
-  if (burst && allBattery) {
-    return allBattery;
+  if (pilot && battery5) {
+    return pilot;
+  }
+
+  if (burst && battery5) {
+    return battery5;
+  }
+
+  if (burst && data.enemy.armdozer.battery === 0) {
+    return burst;
   }
 
   if (battery1) {
@@ -59,12 +61,12 @@ const defenseRoutine: SimpleRoutine = data => {
 };
 
 /**
- * バースト発動 ネオランドーザ
+ * ハードコース ネオランドーザ NPC
  *
  * @return NPC
  */
-export function burstNeoLandozer(): NPC {
+export function hardNeoLandozer(): NPC {
   const armdozer = ArmDozers.find(v => v.id === ArmDozerIdList.NEO_LANDOZER) ?? ArmDozers[0];
-  const pilot = Pilots.find(v => v.id === PilotIds.GAI) ?? Pilots[0];
+  const pilot = Pilots.find(v => v.id === PilotIds.TSUBASA) ?? Pilots[0];
   return new SimpleNPC(armdozer, pilot, attackRoutine, defenseRoutine);
 }
