+++ src/js/npc/hard-shin-braver.js
@@ -1,0 +1,77 @@
+// @flow
+
+import type {NPC} from "./npc";
+import {ArmDozerIdList, ArmDozers, PilotIds, Pilots} from "gbraver-burst-core";
+import type {SimpleRoutine} from "./simple-npc";
+import {SimpleNPC} from "./simple-npc";
+
+/** 0バッテリー */
+const ZERO_BATTERY = {
+  type: 'BATTERY_COMMAND',
+  battery: 0
+};
+
+/**
+ * @override
+ * 攻撃ルーチン
+ */
+const attackRoutine: SimpleRoutine = data => {
+  const burst = data.commands.find(v => v.type === 'BURST_COMMAND');
+  const allBattery = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === data.enemy.armdozer.battery);
+  const allBatteryMinusOne = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === data.enemy.armdozer.battery - 1);
+
+  if (allBattery && burst) {
+    return allBattery;
+  }
+
+  if (allBatteryMinusOne) {
+    return allBatteryMinusOne;
+  }
+
+  return ZERO_BATTERY;
+};
+
+/**
+ * @override
+ * 防御ルーチン
+ */
+const defenseRoutine: SimpleRoutine = data => {
+  const burst = data.commands.find(v => v.type === 'BURST_COMMAND');
+  const pilot = data.commands.find(v => v.type === 'PILOT_SKILL_COMMAND');
+  const battery5 = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === 5);
+  const battery3 = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === 3);
+  const battery1 = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === 1);
+
+  if (data.enemy.armdozer.battery === 0 && burst) {
+    return burst;
+  }
+
+  if (!burst && pilot && battery5) {
+    return battery5;
+  }
+
+  if (pilot) {
+    return pilot;
+  }
+
+  if (burst && battery3) {
+    return battery3;
+  }
+
+  if (battery1) {
+    return battery1;
+  }
+
+  return ZERO_BATTERY;
+};
+
+/**
+ * ハードコース シンブレイバー NPC
+ *
+ * @returns NPC
+ */
+export function hardShinBraver(): NPC {
+  const armdozer = ArmDozers.find(v => v.id === ArmDozerIdList.SHIN_BRAVER) ?? ArmDozers[0];
+  const pilot = Pilots.find(v => v.id === PilotIds.RAITO) ?? Pilots[0];
+  return new SimpleNPC(armdozer, pilot, attackRoutine, defenseRoutine);
+}
