+++ src/js/npc/hard-wing-dozer.js
@@ -1,7 +1,7 @@
 // @flow
 
 import type {NPC} from "./npc";
-import {ArmDozerIdList, ArmDozers, correctPower, PilotIds, Pilots} from "gbraver-burst-core";
+import {ArmDozerIdList, ArmDozers, PilotIds, Pilots} from "gbraver-burst-core";
 import type {SimpleRoutine} from "./simple-npc";
 import {SimpleNPC} from "./simple-npc";
 
@@ -16,17 +16,16 @@ const ZERO_BATTERY = {
  * 攻撃ルーチン
  */
 const attackRoutine: SimpleRoutine = data => {
-  const hasCorrectPower = 0 < correctPower(data.enemy.armdozer.effects);
   const pilot = data.commands.find(v => v.type === 'PILOT_SKILL_COMMAND');
   const battery5 = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === 5);
-  const allBatteryMinusOne = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === data.enemy.armdozer.battery - 1);
+  const allBatteryMinusOne = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === data.enemy.armdozer.battery - 1)
 
-  if (pilot) {
-    return pilot;
+  if (pilot && battery5) {
+    return battery5;
   }
 
-  if (hasCorrectPower && battery5) {
-    return battery5;
+  if (data.enemy.armdozer.battery === 0 && pilot) {
+    return pilot;
   }
 
   if (allBatteryMinusOne) {
@@ -48,20 +47,16 @@ const defenseRoutine: SimpleRoutine = data => {
     return burst;
   }
 
-  if (battery1) {
-    return battery1;
-  }
-
-  return ZERO_BATTERY;
+  return battery1 ?? ZERO_BATTERY;
 };
 
 /**
- * 速攻戦法 ライトニングドーザ NPC
+ * ハードコース ウィングドーザ NPC
  *
- * @returns NPC
+ * @return NPC
  */
-export function swiftAttackLightningDozer(): NPC {
-  const armdozer = ArmDozers.find(v => v.id === ArmDozerIdList.LIGHTNING_DOZER) ?? ArmDozers[0];
-  const pilot = Pilots.find(v => v.id === PilotIds.GAI) ?? Pilots[0];
+export function hardWingDozerNPC(): NPC {
+  const armdozer = ArmDozers.find(v => v.id === ArmDozerIdList.WING_DOZER) ?? ArmDozers[0];
+  const pilot = Pilots.find(v => v.id === PilotIds.SHINYA) ?? Pilots[0];
   return new SimpleNPC(armdozer, pilot, attackRoutine, defenseRoutine);
 }
