+++ src/js/npc/normal-neo-landozer.js
@@ -1,10 +1,9 @@
 // @flow
 
 import type {NPC} from "./npc";
-import {ArmDozerIdList, ArmDozers, PilotIds, Pilots} from "gbraver-burst-core";
+import {ArmDozerIdList, ArmDozers, correctPower, PilotIds, Pilots} from "gbraver-burst-core";
 import type {SimpleRoutine} from "./simple-npc";
 import {SimpleNPC} from "./simple-npc";
-import {canBeatDown} from "./can-beat-down";
 
 /** 0バッテリー */
 const ZERO_BATTERY = {
@@ -20,13 +19,14 @@ const attackRoutine: SimpleRoutine = data => {
   const burst = data.commands.find(v => v.type === 'BURST_COMMAND');
   const allBattery = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === data.enemy.armdozer.battery);
   const allBatteryMinusOne = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === data.enemy.armdozer.battery - 1);
-  const canBeatDownWithAllBattery = canBeatDown(data.enemy, data.enemy.armdozer.battery, data.player, data.player.armdozer.battery);
+  const hasPlusCorrectPower = 0 < correctPower(data.enemy.armdozer.effects);
+  const hasDamaged = data.enemy.armdozer.hp  < data.enemy.armdozer.maxHp;
 
-  if (burst && allBattery) {
-    return allBattery;
+  if (hasDamaged && burst) {
+    return burst;
   }
 
-  if (canBeatDownWithAllBattery && !data.player.armdozer.enableBurst && !data.player.pilot.enableSkill && allBattery) {
+  if (hasPlusCorrectPower && allBattery) {
     return allBattery;
   }
 
@@ -44,35 +44,26 @@ const attackRoutine: SimpleRoutine = data => {
 const defenseRoutine: SimpleRoutine = data => {
   const burst = data.commands.find(v => v.type === 'BURST_COMMAND');
   const allBattery = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === data.enemy.armdozer.battery);
-  const battery3 = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === 3);
   const battery1 = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === 1);
 
-  if (burst && data.enemy.armdozer.battery === 0) {
-    return burst;
-  }
-
-  if (battery3 && data.enemy.armdozer.battery === 5) {
-    return battery3;
+  if (burst && allBattery) {
+    return allBattery;
   }
 
   if (battery1) {
     return battery1;
   }
 
-  if (allBattery) {
-    return allBattery;
-  }
-
   return ZERO_BATTERY;
 };
 
 /**
- * バースト発動 シンブレイバー NPC
+ * ノーマルコース ネオランドーザNPC
  *
- * @returns NPC
+ * @return NPC
  */
-export function burstShinBraver(): NPC {
-  const armdozer = ArmDozers.find(v => v.id === ArmDozerIdList.SHIN_BRAVER) ?? ArmDozers[0];
-  const pilot = Pilots.find(v => v.id === PilotIds.SHINYA) ?? Pilots[0];
+export function normalNeoLandozer(): NPC {
+  const armdozer = ArmDozers.find(v => v.id === ArmDozerIdList.NEO_LANDOZER) ?? ArmDozers[0];
+  const pilot = Pilots.find(v => v.id === PilotIds.GAI) ?? Pilots[0];
   return new SimpleNPC(armdozer, pilot, attackRoutine, defenseRoutine);
 }
