+++ src/js/npc/normal-shin-braver.js
@@ -4,6 +4,7 @@ import type {NPC} from "./npc";
 import {ArmDozerIdList, ArmDozers, PilotIds, Pilots} from "gbraver-burst-core";
 import type {SimpleRoutine} from "./simple-npc";
 import {SimpleNPC} from "./simple-npc";
+import {canBeatDown} from "./can-beat-down";
 
 /** 0バッテリー */
 const ZERO_BATTERY = {
@@ -19,8 +20,13 @@ const attackRoutine: SimpleRoutine = data => {
   const burst = data.commands.find(v => v.type === 'BURST_COMMAND');
   const allBattery = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === data.enemy.armdozer.battery);
   const allBatteryMinusOne = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === data.enemy.armdozer.battery - 1);
+  const canBeatDownWithAllBattery = canBeatDown(data.enemy, data.enemy.armdozer.battery, data.player, data.player.armdozer.battery);
 
-  if (allBattery && burst) {
+  if (burst && allBattery) {
+    return allBattery;
+  }
+
+  if (canBeatDownWithAllBattery && !data.player.armdozer.enableBurst && !data.player.pilot.enableSkill && allBattery) {
     return allBattery;
   }
 
@@ -37,24 +43,15 @@ const attackRoutine: SimpleRoutine = data => {
  */
 const defenseRoutine: SimpleRoutine = data => {
   const burst = data.commands.find(v => v.type === 'BURST_COMMAND');
-  const pilot = data.commands.find(v => v.type === 'PILOT_SKILL_COMMAND');
-  const battery5 = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === 5);
+  const allBattery = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === data.enemy.armdozer.battery);
   const battery3 = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === 3);
   const battery1 = data.commands.find(v => v.type === 'BATTERY_COMMAND' && v.battery === 1);
 
-  if (data.enemy.armdozer.battery === 0 && burst) {
+  if (burst && data.enemy.armdozer.battery === 0) {
     return burst;
   }
 
-  if (!burst && pilot && battery5) {
-    return battery5;
-  }
-
-  if (pilot) {
-    return pilot;
-  }
-
-  if (burst && battery3) {
+  if (battery3 && data.enemy.armdozer.battery === 5) {
     return battery3;
   }
 
@@ -62,16 +59,20 @@ const defenseRoutine: SimpleRoutine = data => {
     return battery1;
   }
 
+  if (allBattery) {
+    return allBattery;
+  }
+
   return ZERO_BATTERY;
 };
 
 /**
- * 防御全振り シンブレイバー NPC
+ * ノーマルコース シンブレイバー NPC
  *
  * @returns NPC
  */
-export function fullDefenseShinBraver(): NPC {
+export function normalShinBraver(): NPC {
   const armdozer = ArmDozers.find(v => v.id === ArmDozerIdList.SHIN_BRAVER) ?? ArmDozers[0];
-  const pilot = Pilots.find(v => v.id === PilotIds.RAITO) ?? Pilots[0];
+  const pilot = Pilots.find(v => v.id === PilotIds.SHINYA) ?? Pilots[0];
   return new SimpleNPC(armdozer, pilot, attackRoutine, defenseRoutine);
 }
