+++ test/js/game/in-progress/npc-battle/simple-npc-battle-course-test.js
@@ -1,15 +1,15 @@
 // @flow
 
 import test from 'ava';
-import {burstNeoLandozer} from '../../../../../src/js/npc/burst-neo-landozer';
+import {normalNeoLandozer} from '../../../../../src/js/npc/normal-neo-landozer';
 import {SimpleNPCBattleCourse} from "../../../../../src/js/game/npc-battle/simple-npc-battle-course";
 import {DefaultStage} from "../../../../../src/js/game/npc-battle/npc-battle-stage";
 
-const stage1 = {caption: ['test stage 1'], npc: burstNeoLandozer()};
-const stage2 = {caption: ['test stage 2'], npc: burstNeoLandozer()};
-const stage3 = {caption: ['test stage 3'], npc: burstNeoLandozer()};
-const stage4 = {caption: ['test stage 4'], npc: burstNeoLandozer()};
-const stage5 = {caption: ['test stage 5'], npc: burstNeoLandozer()};
+const stage1 = {caption: ['test stage 1'], npc: normalNeoLandozer()};
+const stage2 = {caption: ['test stage 2'], npc: normalNeoLandozer()};
+const stage3 = {caption: ['test stage 3'], npc: normalNeoLandozer()};
+const stage4 = {caption: ['test stage 4'], npc: normalNeoLandozer()};
+const stage5 = {caption: ['test stage 5'], npc: normalNeoLandozer()};
 const stages = [stage1, stage2, stage3, stage4, stage5];
 
 test('引数に指定したステージ数をXとすると、stagesのX番目の要素が対応するステージである', t => {
